use as cli
