+++ cli/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+contract-bindings = { path= "./bindings" }
+clap = { version = "4.0.32", features = ["derive"] }
+ethers = { git = "https://github.com/gakonst/ethers-rs", default-features = false, rev = 'ee5e3e52', features = ["abigen"]  }
+tokio = { version = "1.17.0", features = ["macros"] }
+eyre = "0.6.6"
+serde_json = "1.0.91"
+serde = "1.0.152"
+log = { version = "0.4.17" }
+colog = { version = "1.1.0" }
+
+
+
