+++ cli/bindings/src/lib.rs
