+++ cli/bindings/src/test_payment_splitter.rs
