+++ cli/src/lib.rs
@@ -1,0 +1,7 @@
+//! A library for deploying saturn contracts
+//!
+
+/// CLI commands.
+pub mod commands;
+/// Utilities
+pub mod utils;
