--- utility-scripts/Cargo.toml
@@ -1,6 +1,0 @@
-[workspace]
-
-members = [
-    "bindings",
-    "functions",
-]
