--- utility-scripts/functions/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "functions"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-contract-bindings = { path= "../bindings" }
-ethers = { git = "https://github.com/gakonst/ethers-rs", default-features = false, ref = 'ee5e3e52' }
-tokio = { version = "1.17.0", features = ["macros"] }
-eyre = "0.6.6"
-serde_json = "1.0.91"
-serde = "1.0.152"
-log = { version = "0.4.17" }
-colog = { version = "1.1.0" }
